// src/app/interfaces/user.ts
export interface User {
    name: string;
    lastname: string;
    born: Date;
    nickname: string;
    email: string;
    password: string;
    city: string;
    country: string;
    webpage: string;
    joined: Date;
    description: string;
    uid: string;
    photo?: string;
    post?: any;
}